+++ warp/tests/test_tile_reduce.py
@@ -28,7 +28,7 @@ def tile_sum_kernel(input: wp.array2d(dtype=float), output: wp.array(dtype=float
     n = input.shape[1]
     count = int(n / TILE_DIM)
 
-    s = wp.tile_zeros(m=1, n=1, dtype=float)
+    s = wp.tile_zeros(m=1, dtype=float)
 
     for j in range(count):
         a = wp.tile_load(input, i, j, m=1, n=TILE_DIM)
@@ -217,7 +217,7 @@ def tile_reduce_simt_kernel(output: wp.array(dtype=int)):
     s = wp.tile_sum(t)  # sum over block
 
     # update global sum
-    wp.tile_atomic_add(output, 0, 0, s)
+    wp.tile_atomic_add(output, 0, s)
 
 
 def test_tile_reduce_simt(test, device):
@@ -335,11 +335,11 @@ def tile_arange_kernel(out: wp.array2d(dtype=int)):
     d = wp.tile_arange(-1, 16, dtype=int)
     e = wp.tile_arange(17, 0, -1, dtype=int)
 
-    wp.tile_store(out, 0, 0, a)
-    wp.tile_store(out, 1, 0, b)
-    wp.tile_store(out, 2, 0, c)
-    wp.tile_store(out, 3, 0, d)
-    wp.tile_store(out, 4, 0, e)
+    wp.tile_store(out[0], 0, a)
+    wp.tile_store(out[1], 0, b)
+    wp.tile_store(out[2], 0, c)
+    wp.tile_store(out[3], 0, d)
+    wp.tile_store(out[4], 0, e)
 
 
 def test_tile_arange(test, device):